// src/app/shared/models/shipment-headers-info.model.ts
export class ShipmentHeadersInfo {
  clientId: number = 0;
  shipmentId: number = 0;
  orderHeaderId: string = '';
  orderNumber: string = '';
  orderType: string = '';
  department: string = '';
  deliveryId: string = '';
  parentDeliveryId: number = 0;
  lpnCount: number = 0;
  originalShipMethod: string = '';
  originalCarrierCode: string = '';
  originalShipToAddressInfo: AddressInfo = new AddressInfo();
  originalPaytermCode: string = '';
  carrierId: number = 0;
  carrierName: string = '';
  motCode: string = '';
  shipMethod: string = '';
  carrierAccountNo: string = '';
  carrierPayCode: string = '';
  applyFreightFlag: boolean = false;
  shipFlag: boolean = false;
  printLabelFlag: boolean = false;
  shipConfirmFlag: boolean = false;
  voidFlag: boolean = false;
  manualTrackingFlag: boolean = false;
  shipmentStatusFlag: boolean = false;
  saturdayShipFlag: boolean = false;
  eodFlag: boolean = false;
  residentialFlag: boolean = false;
  releasedStatus: boolean = false;
  separateShipFlag: boolean = false;
  largePkgFlag: boolean = false;
  shipmentDate: string = '';
  printDate: string = '';
  shipConfirmDate: string = '';
  shipconfirmFailedDate: string = '';
  shipconfirmFailedFlag: boolean = false;
  shipconfirmFailedMessage: string = '';
  orgId: string = '';
  invOrgId: string = '';
  shipperName: string = '';
  currencyCode: string = '';
  additionalShipmentInfo: AdditionalShipmentInfo = new AdditionalShipmentInfo();
  waybillNumber: string = '';
  bolnumber: string = '';
  reference1: string = '';
  reference2: string = '';
  totalPkgWeight: number = 0;
  numberOfCartons: number = 0;
  maxLinesCount: number = 0;
  pickUpType: string = '';
  packaging: string = '';
  packSlipRequestId: number = 0;
  distributionCenter: string = '';
  bundleId: string = '';
  jobSetName: string = '';
  jobSetPackageName: string = '';
  mbolFlag: string = '';
  mbolFilename: string = '';
  mbolGenForDelivery: string = '';
  batchId: number = 0;
  shipFromAddrInfo: AddressInfo = new AddressInfo();
  shipToAddrInfo: AddressInfo = new AddressInfo();
  billToAddrInfo: AddressInfo = new AddressInfo();
  shipmentChargesInfo: ShipmentChargesInfo = new ShipmentChargesInfo();
  emailNotificationInfo: EmailNotificationInfo = new EmailNotificationInfo();
  printerInfo: PrinterInfo = new PrinterInfo();
  paytermInfo: PaytermInfo = new PaytermInfo();
  customAttributesInfo: CustomAttributesInfo = new CustomAttributesInfo();
  worldEaseFlag: boolean = false;
  consolidationIndex: string = '';
  intlFlag: boolean = false;
  intlImporterInfo: AddressInfo = new AddressInfo();
  intlSoldToInfo: AddressInfo = new AddressInfo();
  intlUscoInfo: IntlUscoInfo = new IntlUscoInfo();
  intlDutiesTaxesInfo: IntlDutiesTaxesInfo = new IntlDutiesTaxesInfo();
  intlCiInfo: IntlCiInfo = new IntlCiInfo();
  intlDocSubmissionType: string = '';
  ciFlag: boolean = false;
  uscoFlag: boolean = false;
  mpsPackageCount: number = 0;
  intlCN22Info: IntlCN22Info = new IntlCN22Info();
  cn22Flag: boolean = false;
  documentInfo: DocumentInfo = new DocumentInfo();

  pickSlipNumber: number = 0;
  rpFedExAccNum: string = '';
  rpFGrdAccNum: string = '';
  rpUpsAccNum: string = '';
  rpDhlAccNum: string = '';
  rpAddressInfo: AddressInfo = new AddressInfo();

  tpFedExAccNum: string = '';
  tpFGrdAccNum: string = '';
  tpUpsAccNum: string = '';
  tpDhlAccNum: string = '';
  tpAddressInfo: AddressInfo = new AddressInfo();

  shipmentLinesInfoList: ShipmentLinesInfo[] = [];
  shipmentPackagesInfoList: ShipmentPackagesInfo[] = [];
  childShipmentHeadersInfoList: ChildShipmentHeadersInfo[] = [];
}